backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1

# PDF text extraction
pypdf2==3.0.1

# Rate limiting
slowapi==0.1.9

# Official Anthropic SDK (more robust than raw HTTP)
anthropic==0.39.0
